v0.4.2: various balance changes

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/glyphs/Swiftness.java
@@ -40,7 +40,7 @@ public class Swiftness extends Armor.Glyph {
 	}
 
 	@Override
-	public int tierSTRAdjust() {
+	public float tierSTRAdjust() {
 		return -1;
 	}
